requirements.txt
sentence-transformers
pygpt4all==1.1.0
chromadb==0.4.15
urllib3==2.0.2
pdfminer.six==20221105
python-dotenv==1.0.0
unstructured==0.6.6
extract-msg==0.41.1
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
langchain==0.0.171
streamlit==1.9.0

